Fix static analysis(cppClean)
Change unnecessary static valuable to member or constant
Remove unused valuables
Remove commented code
Remove redundant include
Remove unused define
Fix incorrect array usage

Bug: 260038109
Bug: 260829813
Test: 1 to 1 Voice/Video/RttCall test with device, atest ImsMediaNativeTests

+++ service/src/com/android/telephony/imsmedia/lib/libimsmedia/core/video/nodes/IVideoRendererNode.cpp
@@ -47,7 +47,7 @@ IVideoRendererNode::IVideoRendererNode(BaseSessionCallback* callback) :
     mHeight = 0;
     mSamplingRate = 0;
     mCvoValue = 0;
-    memset(mConfigBuffer, 0, MAX_CONFIG_INDEX * MAX_CONFIG_LEN * sizeof(uint32_t));
+    memset(mConfigBuffer, 0, MAX_CONFIG_INDEX * MAX_CONFIG_LEN * sizeof(uint8_t));
     memset(mConfigLen, 0, MAX_CONFIG_INDEX * sizeof(uint32_t));
     mDeviceOrientation = 0;
     mFirstFrame = false;
